refactor(cli): extract shared selection UI and helpers from assistants/skills setup
