+++ .codemie/codemie-cli.config.json
@@ -15,13 +15,31 @@
       "name": "epm-cdme",
       "codemieAssistants": [
         {
+          "id": "0cd793d5-c302-4250-9dc0-a86e46749329",
+          "name": "[QA] Checklist Assistant",
+          "slug": "qachecklist-assistant",
+          "description": "QA Checklist Creation Assistant. The main purpose is to generate checklist for testing purpose based on the user story and divided by test type that could be applied to the requirements.",
+          "project": "epm-cdme",
+          "registeredAt": "2026-04-20T13:47:04.094Z",
+          "registrationMode": "agent"
+        },
+        {
+          "id": "4e2ef106-f20c-4506-a638-df88ad7f6022",
+          "name": "sdsds",
+          "slug": "sdsds",
+          "description": "sdsd",
+          "project": "epm-cdme",
+          "registeredAt": "2026-04-20T13:47:04.096Z",
+          "registrationMode": "agent"
+        },
+        {
           "id": "05959338-06de-477d-9cc3-08369f858057",
           "name": "AI/Run FAQ",
           "slug": "codemie-onboarding",
           "description": "This is smart CodeMie assistant which can help you with onboarding process.\nCodeMie can answer to all you questions about capabilities, usage and so on.",
           "project": "codemie",
-          "registeredAt": "2026-03-18T18:38:32.179Z",
-          "registrationMode": "skill"
+          "registeredAt": "2026-04-20T13:47:04.096Z",
+          "registrationMode": "agent"
         },
         {
           "id": "0368dce9-3987-49ac-b12e-41ce45623a20",
@@ -29,8 +47,8 @@
           "slug": "sonarqube-mcp-analyzer",
           "description": "A highly specialized assistant designed to analyze SonarQube reports using SonarQube MCP Server tools. It processes report links, interpreting all available metrics such as the number and types of issues, severities, affected code snippets, coverage details, and more. Serving both direct users and other AI Assistants, it delivers in-depth insights and actionable recommendations on code quality, technical debt, and coverage improvement.",
           "project": "epm-cdme",
-          "registeredAt": "2026-03-18T18:38:32.180Z",
-          "registrationMode": "skill"
+          "registeredAt": "2026-04-20T13:47:04.097Z",
+          "registrationMode": "agent"
         },
         {
           "id": "f14e801a-1e6c-4d2a-ab70-f59795c11a1b",
@@ -38,8 +56,8 @@
           "slug": "brianna",
           "description": "Business Analyst Assistant - expert to work with Jira. Used for creating/getting/managing Jira tickets in EPM-CDME project (Epics, Stories, Tasks, and Bugs). Main role is to analyze requirements from the request, clarify additional questions if necessary, generate requirements with the description structure defined in the prompt and additional details from the request, and create tickets in EPM-CDME project Jira. The Assistant uses Generic Jira tool for Jira tickets creation.",
           "project": "epm-cdme",
-          "registeredAt": "2026-03-18T18:38:32.181Z",
-          "registrationMode": "skill"
+          "registeredAt": "2026-04-20T13:47:04.098Z",
+          "registrationMode": "agent"
         }
       ]
     }
