+++ src/cli/commands/assistants/setup/generators/claude-skill-generator.ts
@@ -8,9 +8,11 @@ import type { Assistant } from 'codemie-sdk';
 /**
  * Get the skills directory path for Claude Code
  */
-function getSkillsDir(): string {
-	const homeDir = os.homedir();
-	return path.join(homeDir, '.claude', 'skills');
+function getSkillsDir(scope: 'global' | 'local' = 'global', workingDir?: string): string {
+	if (scope === 'local' && workingDir) {
+		return path.join(workingDir, '.claude', 'skills');
+	}
+	return path.join(os.homedir(), '.claude', 'skills');
 }
 
 /**
@@ -82,11 +84,11 @@ function createSkillContent(assistant: Assistant): string {
 
 /**
  * Register an assistant as a Claude Code skill
- * Creates: ~/.claude/skills/{slug}/SKILL.md
+ * Creates: ~/.claude/skills/{slug}/SKILL.md (global) or {cwd}/.claude/skills/{slug}/SKILL.md (local)
  */
-export async function registerClaudeSkill(assistant: Assistant): Promise<void> {
+export async function registerClaudeSkill(assistant: Assistant, scope: 'global' | 'local' = 'global', workingDir?: string): Promise<void> {
 	const slug = assistant.slug || assistant.id.toLowerCase().replace(/[^a-z0-9]+/g, '-');
-	const skillsDir = getSkillsDir();
+	const skillsDir = getSkillsDir(scope, workingDir);
 	const skillDir = path.join(skillsDir, slug);
 	const skillFile = path.join(skillDir, 'SKILL.md');
 
@@ -105,10 +107,10 @@ export async function registerClaudeSkill(assistant: Assistant): Promise<void> {
 
 /**
  * Unregister a Claude Code skill
- * Removes: ~/.claude/skills/{slug}/
+ * Removes: ~/.claude/skills/{slug}/ (global) or {cwd}/.claude/skills/{slug}/ (local)
  */
-export async function unregisterClaudeSkill(slug: string): Promise<void> {
-	const skillsDir = getSkillsDir();
+export async function unregisterClaudeSkill(slug: string, scope: 'global' | 'local' = 'global', workingDir?: string): Promise<void> {
+	const skillsDir = getSkillsDir(scope, workingDir);
 	const skillDir = path.join(skillsDir, slug);
 
 	try {
