+++ src/cli/commands/assistants/setup/helpers.ts
@@ -1,11 +1,4 @@
-/**
- * Registration Operations Module
- *
- * Handles registration/unregistration business logic
- */
-
 import chalk from 'chalk';
-import ora from 'ora';
 import type { Assistant, AssistantBase } from 'codemie-sdk';
 import type { CodemieAssistant } from '@/env/types.js';
 import { logger } from '@/utils/logger.js';
@@ -14,78 +7,29 @@ import { registerClaudeSubagent, unregisterClaudeSubagent } from '@/cli/commands
 import { registerClaudeSkill, unregisterClaudeSkill } from '@/cli/commands/assistants/setup/generators/claude-skill-generator.js';
 import type { RegistrationMode } from '@/cli/commands/assistants/setup/manualConfiguration/types.js';
 import { REGISTRATION_MODE } from '@/cli/commands/assistants/setup/manualConfiguration/constants.js';
+import { executeWithSpinner, determineChanges as _determineChanges } from '@/cli/commands/shared/helpers.js';
+
+export { executeWithSpinner };
 
 export interface RegistrationChanges {
   toRegister: Assistant[];
   toUnregister: CodemieAssistant[];
 }
 
-/**
- * Determine which assistants to register and unregister
- */
 export function determineChanges(
   selectedIds: string[],
   allAssistants: (Assistant | AssistantBase)[],
   registeredAssistants: CodemieAssistant[]
 ): RegistrationChanges {
-  const selectedSet = new Set(selectedIds);
-  const registeredIds = new Set(registeredAssistants.map(a => a.id));
-
-  const toRegister = allAssistants.filter(
-    a => selectedSet.has(a.id) && !registeredIds.has(a.id)
-  ) as Assistant[];
-
-  const toUnregister = registeredAssistants.filter(a => !selectedSet.has(a.id));
-
-  return { toRegister, toUnregister };
+  return _determineChanges(selectedIds, allAssistants as Assistant[], registeredAssistants) as RegistrationChanges;
 }
 
-/**
- * Execute assistant operation with spinner
- */
-async function executeWithSpinner<T>(
-  spinnerMessage: string,
-  operation: () => Promise<T>,
-  successMessage: string,
-  errorMessage: string,
-  onError?: (error: unknown) => void
-): Promise<T | null> {
-  const isVerbose = process.env.CODEMIE_DEBUG === 'true';
-  const spinner = ora(spinnerMessage).start();
-
-  try {
-    const result = await operation();
-    if (isVerbose) {
-      spinner.succeed(chalk.green(successMessage));
-    } else {
-      spinner.clear();
-      spinner.stop();
-    }
-    return result;
-  } catch (error) {
-    if (isVerbose) {
-      spinner.fail(chalk.red(errorMessage));
-    } else {
-      spinner.clear();
-      spinner.stop();
-    }
-    if (onError) {
-      onError(error);
-    }
-    return null;
-  }
-}
-
-/**
- * Unregister an assistant
- * Removes both Claude agent and skill files
- */
-export async function unregisterAssistant(assistant: CodemieAssistant): Promise<void> {
+export async function unregisterAssistant(assistant: CodemieAssistant, scope: 'global' | 'local' = 'global', workingDir?: string): Promise<void> {
   await executeWithSpinner(
     MESSAGES.SETUP.SPINNER_UNREGISTERING(chalk.bold(assistant.name)),
     async () => {
-      await unregisterClaudeSubagent(assistant.slug);
-      await unregisterClaudeSkill(assistant.slug);
+      await unregisterClaudeSubagent(assistant.slug, scope, workingDir);
+      await unregisterClaudeSkill(assistant.slug, scope, workingDir);
     },
     MESSAGES.SETUP.SUCCESS_UNREGISTERED(chalk.bold(assistant.name), chalk.cyan(assistant.slug)),
     MESSAGES.SETUP.ERROR_UNREGISTER_FAILED(assistant.name),
@@ -93,13 +37,11 @@ export async function unregisterAssistant(assistant: CodemieAssistant): Promise<
   );
 }
 
-/**
- * Register an assistant with specified registration mode
- * @param mode - 'agent' (Claude agent only) or 'skill' (Claude skill only)
- */
 export async function registerAssistant(
   assistant: Assistant,
-  mode: RegistrationMode = REGISTRATION_MODE.AGENT
+  mode: RegistrationMode = REGISTRATION_MODE.AGENT,
+  scope: 'global' | 'local' = 'global',
+  workingDir?: string
 ): Promise<CodemieAssistant | null> {
   const modeLabel = mode === REGISTRATION_MODE.SKILL ? 'skill' : 'agent';
 
@@ -108,11 +50,11 @@ export async function registerAssistant(
     async () => {
       switch (mode) {
         case REGISTRATION_MODE.AGENT:
-          await registerClaudeSubagent(assistant);
+          await registerClaudeSubagent(assistant, scope, workingDir);
           break;
 
         case REGISTRATION_MODE.SKILL:
-          await registerClaudeSkill(assistant);
+          await registerClaudeSkill(assistant, scope, workingDir);
           break;
       }
 
@@ -134,6 +76,6 @@ export async function registerAssistant(
     description: assistant.description,
     project: assistant.project,
     registeredAt: new Date().toISOString(),
-    registrationMode: mode
+    registrationMode: mode,
   };
 }
