+++ src/cli/commands/assistants/setup/index.ts
@@ -1,15 +1,8 @@
-/**
- * Setup Assistants Command - Orchestration
- *
- * Unified command to view, register, and unregister CodeMie assistants
- */
-
 import { Command } from 'commander';
 import chalk from 'chalk';
 import type { Assistant, AssistantBase } from 'codemie-sdk';
 import { logger } from '@/utils/logger.js';
 import { ConfigLoader } from '@/utils/config.js';
-import { createErrorContext, formatErrorForUser } from '@/utils/errors.js';
 import type { CodemieAssistant } from '@/env/types.js';
 import { MESSAGES, ACTIONS } from '@/cli/commands/assistants/constants.js';
 import { getAuthenticatedClient } from '@/utils/auth.js';
@@ -22,6 +15,8 @@ import type { RegistrationMode } from '@/cli/commands/assistants/setup/manualCon
 import { REGISTRATION_MODE } from '@/cli/commands/assistants/setup/manualConfiguration/constants.js';
 import { displaySummary } from '@/cli/commands/assistants/setup/summary/index.js';
 import { ACTION_TYPE } from '@/cli/commands/assistants/setup/constants.js';
+import { enableVerboseLogging, handleSetupError } from '@/cli/commands/shared/helpers.js';
+import { promptStorageScope } from '@/cli/commands/shared/prompts/storage-scope.js';
 
 export interface SetupCommandOptions {
   profile?: string;
@@ -36,9 +31,6 @@ interface ApplyChangesResult {
   unregistered: CodemieAssistant[];
 }
 
-/**
- * Create assistants setup command
- */
 export function createAssistantsSetupCommand(): Command {
   const command = new Command('setup');
 
@@ -56,168 +48,43 @@ export function createAssistantsSetupCommand(): Command {
       try {
         await setupAssistants(options);
       } catch (error: unknown) {
-        handleError(error);
+        handleSetupError(error, 'setup assistants');
       }
     });
 
   return command;
 }
 
-/**
- * Enable verbose debug logging
- */
-function enableVerboseLogging(): void {
-  process.env.CODEMIE_DEBUG = 'true';
-  const logFilePath = logger.getLogFilePath();
-  if (logFilePath) {
-    console.log(chalk.dim(`Debug logs: ${logFilePath}\n`));
-  }
-}
-
-/**
- * Handle command errors
- */
-function handleError(error: unknown): never {
-  const context = createErrorContext(error);
-  logger.error('Failed to setup assistants', context);
-  console.error(formatErrorForUser(context));
-  process.exit(1);
-}
-
-/**
- * Prompt user to choose where to save assistant configuration.
- * Implemented as a raw-mode TUI (same as promptModeSelection) to avoid
- * stdin state issues after the custom TUI prompts that precede it.
- */
-async function promptStorageScope(): Promise<'global' | 'local'> {
-  const ANSI = {
-    CLEAR_SCREEN: '\x1B[2J\x1B[H',
-    HIDE_CURSOR: '\x1B[?25l',
-    SHOW_CURSOR: '\x1B[?25h',
-  } as const;
-
-  const KEY = {
-    UP: '\x1B[A',
-    DOWN: '\x1B[B',
-    ENTER: '\r',
-    ESC: '\x1B',
-    CTRL_C: '\x03',
-  } as const;
-
-  const choices = ['global', 'local'] as const;
-  let selectedIndex = 0;
-
-  function renderUI(): string {
-    const lines: string[] = [
-      '',
-      `  ${MESSAGES.SETUP.PROMPT_STORAGE_SCOPE}`,
-      '',
-    ];
-
-    choices.forEach((choice, i) => {
-      const marker = i === selectedIndex ? chalk.cyan('●') : chalk.dim('○');
-      const label = choice === 'global'
-        ? `${chalk.cyan('Global')} ${chalk.dim(MESSAGES.SETUP.STORAGE_GLOBAL_LABEL)}`
-        : `${chalk.yellow('Local')} ${chalk.dim(MESSAGES.SETUP.STORAGE_LOCAL_LABEL)}`;
-      lines.push(`  ${marker} ${label}`);
-    });
-
-    lines.push('');
-    lines.push(chalk.dim('  ↑↓ Navigate   Enter Confirm'));
-
-    if (selectedIndex === 1) {
-      lines.push('');
-      lines.push(chalk.dim(`  ${MESSAGES.SETUP.STORAGE_LOCAL_NOTE}`));
-    }
-
-    lines.push('');
-    return lines.join('\n');
-  }
-
-  return new Promise((resolve) => {
-    let keepAliveTimer: NodeJS.Timeout | null = null;
-
-    function cleanup() {
-      if (keepAliveTimer) {
-        clearInterval(keepAliveTimer);
-        keepAliveTimer = null;
-      }
-      process.stdin.setRawMode(false);
-      process.stdin.pause();
-      process.stdin.removeAllListeners('data');
-      process.stdout.write(ANSI.SHOW_CURSOR + ANSI.CLEAR_SCREEN);
-    }
-
-    function stop(choice: 'global' | 'local') {
-      cleanup();
-      resolve(choice);
-    }
-
-    function render() {
-      process.stdout.write(ANSI.CLEAR_SCREEN + ANSI.HIDE_CURSOR + renderUI());
-    }
-
-    process.stdin.setRawMode(true);
-    process.stdin.resume();
-    process.stdin.setEncoding('utf8');
-
-    process.stdin.on('data', (key: string) => {
-      switch (key) {
-        case KEY.UP:
-          selectedIndex = Math.max(0, selectedIndex - 1);
-          render();
-          break;
-        case KEY.DOWN:
-          selectedIndex = Math.min(choices.length - 1, selectedIndex + 1);
-          render();
-          break;
-        case KEY.ENTER:
-          stop(choices[selectedIndex]);
-          break;
-        case KEY.ESC:
-        case KEY.CTRL_C:
-          stop('global');
-          break;
-      }
-    });
-
-    keepAliveTimer = setInterval(() => {}, 60000);
-    render();
-  });
-}
-
-/**
- * Setup assistants - unified list/register/unregister
- */
 async function setupAssistants(options: SetupCommandOptions): Promise<void> {
-  // Load config and profile
-  const config = await ConfigLoader.load();
   const profileName = options.profile || await ConfigLoader.getActiveProfileName() || 'default';
+  const workingDir = process.cwd();
   logger.debug('Setting up assistants', { profileName, options });
 
-  // Get authenticated client and current registrations
-  const client = await getAuthenticatedClient(config);
+  const storageScope = await promptStorageScope({
+    title: MESSAGES.SETUP.PROMPT_STORAGE_SCOPE,
+    localNote: MESSAGES.SETUP.STORAGE_LOCAL_NOTE,
+  });
 
-  // Prompt user to select assistants
+  const config = await ConfigLoader.load(workingDir, { name: profileName });
+  const client = await getAuthenticatedClient(config);
+  const registeredAssistants = await ConfigLoader.loadAssistantsByScope(storageScope, workingDir, profileName);
+  config.codemieAssistants = registeredAssistants;
+
   const { selectedIds, action } = await promptAssistantSelection(config, options, client);
   if (action === ACTIONS.CANCEL) {
     console.log(chalk.dim(MESSAGES.SETUP.NO_CHANGES_MADE));
     return;
   }
 
-  // Fetch full details for selected assistants
   const fetcher = createDataFetcher({ config, client, options });
   const selectedAssistants = await fetcher.fetchAssistantsByIds(selectedIds, []);
 
-  // Configure registration modes (skip if no assistants selected)
   let registrationModes = new Map<string, RegistrationMode>();
 
   if (selectedAssistants.length > 0) {
-    // Loop to allow going back
     let configurationComplete = false;
 
     while (!configurationComplete) {
-      // Step 1: Show mode selection screen (Subagents, Skills, Manual)
       const { choice, cancelled, back } = await promptModeSelection();
 
       if (cancelled) {
@@ -226,23 +93,20 @@ async function setupAssistants(options: SetupCommandOptions): Promise<void> {
       }
 
       if (back) {
-        // Go back to selection screen
         return setupAssistants(options);
       }
 
-      // Step 2: Handle choice
-      if (choice === CONFIGURATION_CHOICE.SUBAGENTS) { // Bulk register all as agents
+      if (choice === CONFIGURATION_CHOICE.SUBAGENTS) {
         for (const assistant of selectedAssistants) {
           registrationModes.set(assistant.id, REGISTRATION_MODE.AGENT);
         }
         configurationComplete = true;
-      } else if (choice === CONFIGURATION_CHOICE.SKILLS) { // Bulk register all as skills
+      } else if (choice === CONFIGURATION_CHOICE.SKILLS) {
         for (const assistant of selectedAssistants) {
           registrationModes.set(assistant.id, REGISTRATION_MODE.SKILL);
         }
         configurationComplete = true;
-      } else { // Manual configuration - show individual configuration screen
-        const registeredAssistants = config.codemieAssistants || [];
+      } else {
         const registeredIds = new Set(registeredAssistants.map(a => a.id));
 
         const { registrationModes: modes, action: configAction } = await promptManualConfiguration(
@@ -257,7 +121,6 @@ async function setupAssistants(options: SetupCommandOptions): Promise<void> {
         }
 
         if (configAction === ACTION_TYPE.BACK) {
-          // Go back to configuration mode selection
           continue;
         }
 
@@ -267,28 +130,22 @@ async function setupAssistants(options: SetupCommandOptions): Promise<void> {
     }
   }
 
-  // Prompt for storage scope before making any changes
-  const storageScope = await promptStorageScope();
-
-  // Apply changes and get summary data
   const { newRegistrations, registered, unregistered } = await applyChanges(
     selectedIds,
     selectedAssistants,
-    config.codemieAssistants || [],
-    registrationModes
+    registeredAssistants,
+    registrationModes,
+    storageScope,
+    workingDir
   );
 
-  // Always reflect new state in config for display purposes
   config.codemieAssistants = newRegistrations;
 
-  // Skip saving (and showing configLocation) when nothing changed
   if (registered.length === 0 && unregistered.length === 0) {
     displaySummary(registered, unregistered, profileName, config);
     return;
   }
 
-  // Save to the appropriate config location
-  const workingDir = process.cwd();
   let configLocation: string;
 
   if (storageScope === 'local') {
@@ -299,21 +156,17 @@ async function setupAssistants(options: SetupCommandOptions): Promise<void> {
     configLocation = `global (~/.codemie/codemie-cli.config.json)`;
   }
 
-  // Display summary
   displaySummary(registered, unregistered, profileName, config, configLocation);
 }
 
-/**
- * Apply registration changes
- * Returns new registrations and lists of what changed
- */
 async function applyChanges(
   selectedIds: string[],
   allAssistants: (Assistant | AssistantBase)[],
   registeredAssistants: CodemieAssistant[],
-  registrationModes: Map<string, RegistrationMode>
+  registrationModes: Map<string, RegistrationMode>,
+  scope: 'global' | 'local' = 'global',
+  workingDir?: string
 ): Promise<ApplyChangesResult> {
-  // Determine what needs to change
   const { toRegister, toUnregister } = determineChanges(selectedIds, allAssistants, registeredAssistants);
   const selectedSet = new Set(selectedIds);
   const toReregister = registeredAssistants.filter(a => selectedSet.has(a.id));
@@ -323,13 +176,10 @@ async function applyChanges(
     return { newRegistrations: registeredAssistants, registered: [], unregistered: [] };
   }
 
-  // Unregister: both removed assistants and those needing re-registration
-  const toUnregisterAll = [...toUnregister, ...toReregister];
-  for (const assistant of toUnregisterAll) {
-    await unregisterAssistant(assistant);
+  for (const assistant of [...toUnregister, ...toReregister]) {
+    await unregisterAssistant(assistant, scope, workingDir);
   }
 
-  // Register all selected assistants with their configured modes
   const newRegistrations: CodemieAssistant[] = [];
   const allToRegister = [...toRegister, ...toReregister];
 
@@ -338,37 +188,29 @@ async function applyChanges(
     if (!fullAssistant) continue;
 
     const mode = registrationModes.get(fullAssistant.id) || REGISTRATION_MODE.AGENT;
-    const registered = await registerAssistant(fullAssistant, mode);
+    const registered = await registerAssistant(fullAssistant, mode, scope, workingDir);
     if (registered) {
       newRegistrations.push(registered);
     }
   }
 
-  // Return results for summary
   return {
     newRegistrations,
     registered: [...toRegister, ...getFullAssistants(toReregister, allAssistants)],
-    unregistered: toUnregister
+    unregistered: toUnregister,
   };
 }
 
-/**
- * Get full assistant details from the list
- */
 function getFullAssistant(
   assistant: Assistant | CodemieAssistant,
   allAssistants: (Assistant | AssistantBase)[]
 ): Assistant | null {
-  // Check if it's already a full Assistant (has registeredAt = it's a CodemieAssistant)
   if ('registeredAt' in assistant) {
     return allAssistants.find(a => a.id === assistant.id) as Assistant || null;
   }
   return assistant as Assistant;
 }
 
-/**
- * Get full assistant details for multiple assistants
- */
 function getFullAssistants(
   assistants: CodemieAssistant[],
   allAssistants: (Assistant | AssistantBase)[]
