+++ src/cli/commands/assistants/setup/selection/ui.ts
@@ -1,26 +1,25 @@
-/**
- * UI Rendering
- *
- * All UI building and rendering functions
- */
-
 import chalk from 'chalk';
-import type { PanelState, SelectionState } from './types.js';
-import { ANSI, BOX, SYMBOL, TEXT, CONFIG, PAGINATION_CONTROL, type PanelId, type PaginationControl } from './constants.js';
+import type { SelectionState, PanelState } from './types.js';
+import { TEXT, CONFIG } from './constants.js';
 import { COLOR } from '../constants.js';
+import { ANSI, SYMBOL } from '@/cli/commands/shared/selection/constants.js';
+import {
+  buildTopLine,
+  buildPanelHeader,
+  buildSearchInput,
+  buildPaginationControls,
+  buildButtons,
+  buildInstructions,
+} from '@/cli/commands/shared/selection/ui.js';
 import { REGISTRATION_MODE } from '../manualConfiguration/constants.js';
 
-/**
- * Render the complete UI
- */
 export function renderUI(state: SelectionState, cursorIndex: number): string {
   const activePanel = state.panels.find(p => p.id === state.activePanelId)!;
 
-  let output = '';
+  let output = ANSI.CURSOR_HOME_CLEAR;
 
-  output += ANSI.CURSOR_HOME_CLEAR;
   output += buildTopLine();
-  output += buildPanelHeader(state.panels, state.activePanelId);
+  output += buildPanelHeader(state.panels, state.activePanelId, TEXT.LABEL);
   output += buildCount(activePanel);
   output += buildSearchInput(state.searchQuery, state.isSearchFocused);
   output += buildAssistantsList(state, activePanel, cursorIndex);
@@ -31,79 +30,16 @@ export function renderUI(state: SelectionState, cursorIndex: number): string {
   return output;
 }
 
-/**
- * Build top line spanning terminal width
- */
-function buildTopLine(): string {
-  const width = process.stdout.columns || 80;
-  const line = BOX.HORIZONTAL.repeat(width);
-  return chalk.rgb(COLOR.PURPLE.r, COLOR.PURPLE.g, COLOR.PURPLE.b)(line) + '\n';
-}
-
-/**
- * Build panel navigation header with hint
- */
-function buildPanelHeader(panels: PanelState[], activeId: PanelId): string {
-  const panelStrings = panels.map(panel => {
-    if (panel.id === activeId) {
-      return chalk.bgRgb(COLOR.PURPLE.r, COLOR.PURPLE.g, COLOR.PURPLE.b).black(` ${panel.label} `);
-    }
-    return chalk.white(panel.label);
-  });
-
-  const label = chalk.rgb(COLOR.PURPLE.r, COLOR.PURPLE.g, COLOR.PURPLE.b).bold(TEXT.LABEL);
-  const panelsLine = panelStrings.join('   ');
-  const hint = chalk.dim(TEXT.TAB_HINT);
-
-  return `${label}   ${panelsLine}   ${hint}\n\n`;
-}
-
-/**
- * Build count above search box
- */
 function buildCount(activePanel: PanelState): string {
   if (activePanel.filteredData.length === 0) {
     return chalk.dim('0 assistants total, Page 1 of 1') + '\n';
   }
 
-  const currentPage = activePanel.currentPage + 1; // Display as 1-indexed
+  const currentPage = activePanel.currentPage + 1;
   const totalPages = activePanel.totalPages;
   const totalAssistants = activePanel.totalItems;
 
   return chalk.dim(`${totalAssistants} assistants total, Page ${currentPage} of ${totalPages}`) + '\n';
-}
-
-/**
- * Build search input display with box
- */
-function buildSearchInput(query: string, isFocused: boolean): string {
-  const width = process.stdout.columns || 80;
-  const innerWidth = width - 2;
-
-  const prefix = ` ${SYMBOL.SEARCH_ICON} `;
-  const contentText = query
-    ? (isFocused ? query + SYMBOL.CURSOR : query)
-    : TEXT.SEARCH_PLACEHOLDER;
-  const visualLength = prefix.length + contentText.length;
-  const paddingNeeded = innerWidth - visualLength;
-
-  const styledCursor = chalk.rgb(COLOR.PURPLE.r, COLOR.PURPLE.g, COLOR.PURPLE.b)(SYMBOL.CURSOR);
-  const styledPlaceholder = chalk.dim(TEXT.SEARCH_PLACEHOLDER);
-  const displayText = query
-    ? (isFocused ? query + styledCursor : query)
-    : styledPlaceholder;
-
-  const contentLine = prefix + displayText + ' '.repeat(Math.max(0, paddingNeeded));
-
-  const borderColor = isFocused
-    ? chalk.rgb(COLOR.PURPLE.r, COLOR.PURPLE.g, COLOR.PURPLE.b)
-    : chalk.white;
-
-  let output = '';
-  output += borderColor(BOX.TOP_LEFT + BOX.HORIZONTAL.repeat(innerWidth) + BOX.TOP_RIGHT) + '\n';
-  output += borderColor(BOX.VERTICAL) + contentLine + borderColor(BOX.VERTICAL) + '\n';
-  output += borderColor(BOX.BOTTOM_LEFT + BOX.HORIZONTAL.repeat(innerWidth) + BOX.BOTTOM_RIGHT) + '\n\n';
-  return output;
 }
 
 function buildAssistantsList(
@@ -126,16 +62,13 @@ function buildAssistantsList(
     return chalk.yellow(TEXT.NO_ASSISTANTS + '\n');
   }
 
-  const displayAssistants = activePanel.filteredData;
-
-  // Create a map for quick lookup of registration modes
   const registeredMap = new Map(
     registeredAssistants.map(a => [a.id, a.registrationMode || REGISTRATION_MODE.AGENT])
   );
 
   let output = '';
 
-  displayAssistants.forEach((assistant, index) => {
+  activePanel.filteredData.forEach((assistant, index) => {
     const isSelected = selectedIds.has(assistant.id);
     const isCursor = index === cursorIndex && !isSearchFocused && isPaginationFocused === null && !areNavigationButtonsFocused;
 
@@ -185,75 +118,4 @@ function buildAssistantsList(
   });
 
   return output;
-}
-
-function buildPaginationControls(
-  activePanel: PanelState,
-  isPaginationFocused: PaginationControl | null,
-  isSearchFocused: boolean
-): string {
-  const totalPages = activePanel.totalPages;
-
-  if (totalPages <= 1) {
-    return '';
-  }
-
-  const currentPage = activePanel.currentPage;
-  const hasPrev = currentPage > 0;
-  const hasNext = currentPage < totalPages - 1;
-
-  const prevLabel = hasPrev ? '[< Prev]' : '[< Prev]';
-  const prevCursor = isPaginationFocused === PAGINATION_CONTROL.PREV && !isSearchFocused
-    ? chalk.rgb(COLOR.PURPLE.r, COLOR.PURPLE.g, COLOR.PURPLE.b)(SYMBOL.CURSOR_INDICATOR)
-    : '  ';
-  const prevText = isPaginationFocused === PAGINATION_CONTROL.PREV && !isSearchFocused
-    ? chalk.rgb(COLOR.PURPLE.r, COLOR.PURPLE.g, COLOR.PURPLE.b).bold(prevLabel)
-    : hasPrev
-      ? chalk.white(prevLabel)
-      : chalk.dim(prevLabel);
-
-  const nextLabel = hasNext ? '[Next >]' : '[Next >]';
-  const nextCursor = isPaginationFocused === PAGINATION_CONTROL.NEXT && !isSearchFocused
-    ? chalk.rgb(COLOR.PURPLE.r, COLOR.PURPLE.g, COLOR.PURPLE.b)(SYMBOL.CURSOR_INDICATOR)
-    : '  ';
-  const nextText = isPaginationFocused === PAGINATION_CONTROL.NEXT && !isSearchFocused
-    ? chalk.rgb(COLOR.PURPLE.r, COLOR.PURPLE.g, COLOR.PURPLE.b).bold(nextLabel)
-    : hasNext
-      ? chalk.white(nextLabel)
-      : chalk.dim(nextLabel);
-
-  const label = chalk.dim('Switch page:');
-  return `${label} ${prevCursor}${prevText}    ${nextCursor}${nextText}\n\n`;
-}
-
-/**
- * Build buttons (Continue / Cancel)
- */
-function buildButtons(state: SelectionState): string {
-  const { areNavigationButtonsFocused, focusedButton, isSearchFocused, isPaginationFocused } = state;
-  const buttonsActive = areNavigationButtonsFocused && !isSearchFocused && isPaginationFocused === null;
-
-  const continueButton = buttonsActive && focusedButton === 'continue'
-    ? chalk.bgRgb(COLOR.PURPLE.r, COLOR.PURPLE.g, COLOR.PURPLE.b).black(` ${TEXT.CONTINUE_BUTTON} `)
-    : chalk.dim(`[${TEXT.CONTINUE_BUTTON}]`);
-
-  const cancelButton = buttonsActive && focusedButton === 'cancel'
-    ? chalk.bgRgb(COLOR.PURPLE.r, COLOR.PURPLE.g, COLOR.PURPLE.b).black(` ${TEXT.CANCEL_BUTTON} `)
-    : chalk.dim(`[${TEXT.CANCEL_BUTTON}]`);
-
-  return `  ${continueButton}  ${cancelButton}\n\n`;
-}
-
-/**
- * Build keyboard instructions
- */
-function buildInstructions(activePanel: PanelState): string {
-  const totalPages = activePanel.totalPages;
-  const hasMultiplePages = totalPages > 1;
-
-  const instructionsText = hasMultiplePages
-    ? TEXT.INSTRUCTIONS_WITH_PAGINATION
-    : TEXT.INSTRUCTIONS;
-
-  return chalk.dim(instructionsText + '\n');
 }
