+++ src/cli/commands/skills/setup/index.ts
@@ -1,19 +1,13 @@
-/**
- * Setup Skills Command
- *
- * Fetches platform skills and registers them as Claude Code skills
- */
-
 import { Command } from 'commander';
 import chalk from 'chalk';
-import { logger } from '@/utils/logger.js';
 import { ConfigLoader } from '@/utils/config.js';
-import { createErrorContext, formatErrorForUser } from '@/utils/errors.js';
 import { getAuthenticatedClient } from '@/utils/auth.js';
 import { createSkillDataFetcher } from './data.js';
 import { promptSkillSelection } from './selection/index.js';
 import { determineChanges, registerSkill, unregisterSkill } from './helpers.js';
 import { ACTION_TYPE } from './constants.js';
+import { enableVerboseLogging, handleSetupError } from '@/cli/commands/shared/helpers.js';
+import { promptStorageScope } from '@/cli/commands/shared/prompts/storage-scope.js';
 import type { CodemieSkill } from '@/env/types.js';
 
 export type { CodemieSkill };
@@ -27,20 +21,13 @@ export function createSkillsSetupCommand(): Command {
     .option('-v, --verbose', 'Enable verbose debug output')
     .action(async (options: { profile?: string; verbose?: boolean }) => {
       if (options.verbose) {
-        process.env.CODEMIE_DEBUG = 'true';
-        const logFilePath = logger.getLogFilePath();
-        if (logFilePath) {
-          console.log(chalk.dim(`Debug logs: ${logFilePath}\n`));
-        }
+        enableVerboseLogging();
       }
 
       try {
         await setupSkills(options);
       } catch (error: unknown) {
-        const context = createErrorContext(error);
-        logger.error('Failed to setup skills', context);
-        console.error(formatErrorForUser(context));
-        process.exit(1);
+        handleSetupError(error, 'setup skills');
       }
     });
 
@@ -99,24 +86,24 @@ async function showDisclaimer(): Promise<boolean> {
 }
 
 async function setupSkills(options: { profile?: string }): Promise<void> {
-  // Resolve profile name first so both load and save use the same profile.
-  // Without this, ConfigLoader.load() may pull a different global activeProfile
-  // than the local config's activeProfile, causing codemieSkills to be invisible.
   const profileName = options.profile ?? await ConfigLoader.getActiveProfileName() ?? 'default';
-  const config = await ConfigLoader.load(process.cwd(), { name: profileName });
-  const client = await getAuthenticatedClient(config);
+  const workingDir = process.cwd();
 
-  // Show disclaimer before skill selection
   const proceed = await showDisclaimer();
   if (!proceed) {
     console.log(chalk.dim('\nNo changes made.\n'));
     return;
   }
 
-  // Get currently registered skills
-  const registeredSkills: CodemieSkill[] = config.codemieSkills || [];
+  const storageScope = await promptStorageScope({
+    title: 'Where would you like to save skills configuration?',
+    localNote: 'Project-scoped skills will override global ones for this repository.',
+  });
 
-  // Show interactive selection UI
+  const config = await ConfigLoader.load(workingDir, { name: profileName });
+  const client = await getAuthenticatedClient(config);
+  const registeredSkills: CodemieSkill[] = await ConfigLoader.loadSkillsByScope(storageScope, workingDir, profileName);
+
   const { selectedIds, action } = await promptSkillSelection(registeredSkills, client);
 
   if (action === ACTION_TYPE.CANCEL) {
@@ -124,11 +111,9 @@ async function setupSkills(options: { profile?: string }): Promise<void> {
     return;
   }
 
-  // Fetch full details for selected skills
   const fetcher = createSkillDataFetcher({ client, registeredSkills });
   const selectedSkills = await fetcher.fetchSkillsByIds(selectedIds, registeredSkills);
 
-  // Determine changes
   const { toRegister, toUnregister } = determineChanges(selectedIds, selectedSkills, registeredSkills);
 
   if (toRegister.length === 0 && toUnregister.length === 0) {
@@ -136,32 +121,35 @@ async function setupSkills(options: { profile?: string }): Promise<void> {
     return;
   }
 
-  // Unregister removed skills
   for (const skill of toUnregister) {
-    await unregisterSkill(skill);
+    await unregisterSkill(skill, storageScope, workingDir);
   }
 
-  // Register new skills
   const newlyRegistered: CodemieSkill[] = [];
   for (const skill of toRegister) {
     const detail = await fetcher.fetchSkillById(skill.id);
-    const registered = await registerSkill(detail);
+    const registered = await registerSkill(detail, storageScope, workingDir);
     if (registered) {
       newlyRegistered.push(registered);
     }
   }
 
-  // Build updated skills list
   const updatedSkills: CodemieSkill[] = [
     ...registeredSkills.filter(s => selectedIds.includes(s.id)),
     ...newlyRegistered,
   ];
 
-  // Save to config
-  config.codemieSkills = updatedSkills;
-  await ConfigLoader.saveProfile(profileName, config);
+  let configLocation: string;
 
-  // Summary
+  if (storageScope === 'local') {
+    await ConfigLoader.saveSkillsToProjectConfig(workingDir, profileName, updatedSkills);
+    configLocation = `${workingDir}/.codemie/codemie-cli.config.json`;
+  } else {
+    config.codemieSkills = updatedSkills;
+    await ConfigLoader.saveProfile(profileName, config);
+    configLocation = `global (~/.codemie/codemie-cli.config.json)`;
+  }
+
   console.log('');
   if (newlyRegistered.length > 0) {
     console.log(chalk.green(`✓ Registered ${newlyRegistered.length} skill(s)`));
@@ -169,5 +157,6 @@ async function setupSkills(options: { profile?: string }): Promise<void> {
   if (toUnregister.length > 0) {
     console.log(chalk.yellow(`○ Unregistered ${toUnregister.length} skill(s)`));
   }
-  console.log(chalk.dim('\nSkills are available in Claude Code as /skill-name commands.\n'));
+  console.log(chalk.dim(`\nSkills saved to: ${configLocation}`));
+  console.log(chalk.dim('Skills are available in Claude Code as /skill-name commands.\n'));
 }
