+++ src/cli/commands/skills/setup/selection/index.ts
@@ -1,9 +1,3 @@
-/**
- * Selection UI Module
- *
- * Main orchestration for skill selection with panel interface
- */
-
 import type { CodeMieClient } from 'codemie-sdk';
 import type { CodemieSkill } from '@/env/types.js';
 import type { SelectionState } from './types.js';
@@ -12,6 +6,7 @@ import { ACTION_TYPE, type ActionType } from '../constants.js';
 import { createSkillDataFetcher } from '../data.js';
 import { createInteractivePrompt, type InteractivePrompt } from './interactive-prompt.js';
 import { createActionHandlers } from './actions.js';
+import { renderUI } from './ui.js';
 import { logger } from '@/utils/logger.js';
 import ora from 'ora';
 
@@ -30,12 +25,8 @@ const DEFAULT_PANEL_PARAMS = {
   currentPage: 0,
   totalItems: 0,
   totalPages: 0,
-}
+};
 
-/**
- * Initialize state with 3 panels (Registered, Project, Marketplace)
- * Smart default: Project tab if no registered skills exist
- */
 function initializeState(registeredSkills: CodemieSkill[]): SelectionState {
   const registeredIds = new Set(registeredSkills.map(s => s.id));
   const hasRegisteredSkills = registeredIds.size > 0;
@@ -60,8 +51,8 @@ function initializeState(registeredSkills: CodemieSkill[]): SelectionState {
       {
         id: PANEL_ID.MARKETPLACE,
         label: 'Marketplace',
-        ...DEFAULT_PANEL_PARAMS
-      }
+        ...DEFAULT_PANEL_PARAMS,
+      },
     ],
     activePanelId: defaultPanelId,
     searchQuery: '',
@@ -71,32 +62,27 @@ function initializeState(registeredSkills: CodemieSkill[]): SelectionState {
     isSearchFocused: false,
     isPaginationFocused: null,
     areNavigationButtonsFocused: false,
-    focusedButton: 'continue'
+    focusedButton: 'continue',
   };
 }
 
-/**
- * Prompt user to select skills with panel interface
- */
 export async function promptSkillSelection(
   registeredSkills: CodemieSkill[],
   client: CodeMieClient
 ): Promise<{ selectedIds: string[]; action: ActionType }> {
   const state = initializeState(registeredSkills);
-  const fetcher = createSkillDataFetcher({
-    client,
-    registeredSkills
-  });
+  const fetcher = createSkillDataFetcher({ client, registeredSkills });
 
   let prompt: InteractivePrompt | null = null;
   let isCancelled = false;
 
   const actionHandlers = createActionHandlers({
     state,
-    fetcher,
+    fetchItems: (params) => fetcher.fetchSkills(params),
+    entityLabel: 'Skill',
     prompt: () => prompt,
     setPrompt: (p) => { prompt = p; },
-    setCancelled: (cancelled) => { isCancelled = cancelled; }
+    setCancelled: (cancelled) => { isCancelled = cancelled; },
   });
 
   const spinner = ora('Loading skills...').start();
@@ -109,7 +95,7 @@ export async function promptSkillSelection(
     const result = await fetcher.fetchSkills({
       scope: state.activePanelId,
       searchQuery: state.searchQuery,
-      page: 0
+      page: 0,
     });
     activePanel.data = result.data;
     activePanel.filteredData = result.data;
@@ -123,22 +109,19 @@ export async function promptSkillSelection(
     activePanel.totalPages = 0;
   }
 
-  // Clear spinner output before starting interactive mode
   process.stdout.write(ANSI.CLEAR_LINE_ABOVE);
 
   prompt = createInteractivePrompt({
     state,
-    actions: actionHandlers
+    actions: actionHandlers,
+    renderFn: (s, cursor) => renderUI(s as SelectionState, cursor),
   });
 
   await prompt.start();
 
   if (isCancelled) {
     logger.debug('[SkillSelection] Selection cancelled');
-    return {
-      selectedIds: [],
-      action: ACTION_TYPE.CANCEL
-    };
+    return { selectedIds: [], action: ACTION_TYPE.CANCEL };
   }
 
   const selectedIdsArray = Array.from(state.selectedIds);
@@ -146,11 +129,8 @@ export async function promptSkillSelection(
     totalSelected: selectedIdsArray.length,
     selectedIds: selectedIdsArray,
     registeredCount: state.registeredIds.size,
-    registeredIds: Array.from(state.registeredIds)
+    registeredIds: Array.from(state.registeredIds),
   });
 
-  return {
-    selectedIds: selectedIdsArray,
-    action: ACTION_TYPE.UPDATE
-  };
+  return { selectedIds: selectedIdsArray, action: ACTION_TYPE.UPDATE };
 }
