+++ src/cli/commands/skills/setup/selection/types.ts
@@ -1,31 +1,15 @@
 import type { SkillListItem, SkillDetail } from 'codemie-sdk';
-import type { PanelId, PaginationControl } from './constants.js';
 import type { CodemieSkill } from '@/env/types.js';
+import type { BaseSelectionState, BasePanelState } from '@/cli/commands/shared/selection/types.js';
 
-export type ButtonType = 'continue' | 'cancel';
+export type { ButtonType } from '@/cli/commands/shared/selection/types.js';
 
-export interface SelectionState {
-  panels: PanelState[];
-  activePanelId: PanelId;
-  searchQuery: string;
-  selectedIds: Set<string>;
-  registeredIds: Set<string>;
-  registeredSkills: CodemieSkill[];
-  isSearchFocused: boolean;
-  isPaginationFocused: PaginationControl | null;
-  areNavigationButtonsFocused: boolean;
-  focusedButton: ButtonType;
-}
-
-export interface PanelState {
-  id: PanelId;
-  label: string;
-  isActive: boolean;
+export interface PanelState extends BasePanelState {
   data: (SkillListItem | SkillDetail)[] | null;
   filteredData: (SkillListItem | SkillDetail)[];
-  isFetching: boolean;
-  error: string | null;
-  currentPage: number;
-  totalItems: number;
-  totalPages: number;
+}
+
+export interface SelectionState extends BaseSelectionState {
+  panels: PanelState[];
+  registeredSkills: CodemieSkill[];
 }
